feat: add functionality to edit/update a movie

+++ internals/models/Movie.go
@@ -3,12 +3,12 @@ package models
 import "time"
 
 type Movie struct {
-	ID          int64     `json:"id"`
+	ID          int       `json:"id"`
 	Title       string    `json:"title"`
 	Description string    `json:"description"`
 	ReleaseDate time.Time `json:"release_date"`
 	MPAARating  string    `json:"mpaa_rating"`
-	Runtime     int64     `json:"runtime"`
+	Runtime     int       `json:"runtime"`
 	Image       string    `json:"image"`
 	CreatedAt   time.Time `json:"-"`
 	UpdatedAt   time.Time `json:"-"`
